requirements.txt
python-dotenv
pandas
numpy
setuptools
pymongo
certifi
pymongo[srv]
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart


# -e .